Create new folder for DomainInfo and move files into it.
- Also starts development of Renew Section
