+++ src/app/tabs/newAdmin/components/index.js
@@ -1,8 +1,6 @@
 export { default as AdminTabComponent } from './AdminTabComponent';
-export { default as DomainInfoComponent } from './DomainInfoComponent';
 export { default as AddressesComponent } from './AddressesComponent';
 export { default as SubdomainsComponent } from './SubdomainsComponent';
 export { default as ResolverComponent } from './ResolverComponent';
 export { default as ReverseComponent } from './ReverseComponent';
 export { default as LeftNavComponent } from './LetftNavComponent';
-export { default as RenewButtonComponent } from './RenewButtonComponent';
