+++ src/app/tabs/newAdmin/containers/index.js
@@ -1,9 +1,6 @@
 export { default as AdminTabContainer } from './AdminTabContainer';
-export { default as DomainInfoContainer } from './DomainInfoContainer';
 export { default as AddressesContainer } from './AddressesContainer';
 export { default as SubdomainsContainer } from './SubdomainsContainer';
 export { default as ResolverContainer } from './ResolverContainer';
 export { default as ReverseContainer } from './ReverseContainer';
 export { default as LeftNavContainer } from './LeftNavContainer';
-export { default as TransferAddressContainer } from './TransferAddressContainer';
-export { default as RenewButtonContainer } from './RenewButtonContainer';
