+++ src/app/tabs/newAdmin/domainInfo/components/DomainInfoComponent.test.js
