+++ src/app/tabs/newAdmin/domainInfo/components/index.js
@@ -1,0 +1,3 @@
+export { default as DomainInfoComponent } from './DomainInfoComponent';
+export { default as RenewButtonComponent } from './RenewButtonComponent';
+export { default as RenewDomainComponent } from './RenewDomainComponent';
