+++ src/app/tabs/newAdmin/domainInfo/containers/index.js
@@ -1,0 +1,4 @@
+export { default as DomainInfoContainer } from './DomainInfoContainer';
+export { default as TransferAddressContainer } from './TransferAddressContainer';
+export { default as RenewButtonContainer } from './RenewButtonContainer';
+export { default as RenewDomainContainer } from './RenewDomainContainer';
