+++ src/app/tabs/newAdmin/domainInfo/types.js
@@ -1,0 +1,11 @@
+export const REQUEST_TRANSFER_DOMAIN = 'REQUEST_TRANSFER_DOMAIN';
+export const RECEIVE_TRANSFER_DOMAIN = 'RECEIVE_TRANSFER_DOMAIN';
+export const ERROR_TRANSFER_DOMAIN = 'ERROR_TRANSFER_DOMAIN';
+export const HANDLE_ERROR_CLOSE = 'HANDLE_ERROR_CLOSE';
+export const HANDLE_SUCCESS_CLOSE = 'HANDLE_SUCCESS_CLOSE';
+
+export const RENEW_DOMAIN_CHECK_SUBDOMAIN = 'RENEW_DOMAIN_CHECK_SUBDOMAIN';
+export const REQUEST_DOMAIN_EXPIRATION_TIME = 'REQUEST_DOMAIN_EXPIRATION_TIME';
+export const RECIEVE_DOMAIN_EXPIRATION_TIME = 'RECIEVE_DOMAIN_EXPIRATION_TIME';
+export const ERROR_DOMAIN_EXIPRATION_TIME = 'ERROR_DOMAIN_EXIPRATION_TIME';
+export const TOGGLE_RENEW_PANEL = 'TOGGLE_RENEW_PANEL';
