+++ src/assets/css/sass/_typography.scss
@@ -37,3 +37,8 @@ p.lead {
 label {
   color: $gray;
 }
+
+p.explanation {
+  font-size: .8em !important;
+  color: $gray;
+}
