+++ src/assets/css/sass/admin/domainInfo.scss
@@ -1,4 +1,10 @@
 .domainInfo {
+  @import 'renew';
+
+  button {
+    padding:10px 20px !important;
+    margin-left: 10px;
+  }
   .domain {
     font-size: 25px;
 
@@ -11,8 +17,6 @@
     text-align: right;
 
     button {
-      padding:10px 20px !important;
-      margin-left: 10px;
       border-radius: 15px;
     }
   }
