+++ src/assets/css/sass/admin/renew.scss
@@ -1,0 +1,10 @@
+.renewDomian {
+  text-align: center;
+  
+  .col-md-3, .offset-md-3 {
+    width: 100% !important;
+    margin-left: 0 !important;
+  }
+
+  
+}
